+++ packages/embed/src/js/biometric-kyc.js
@@ -2,6 +2,7 @@ import JSZip from 'jszip';
 import validate from 'validate.js';
 import '@smileid/web-components/end-user-consent';
 import { version as sdkVersion } from '../../package.json';
+import { getMetadata } from './metadata';
 
 (function biometricKyc() {
   'use strict';
@@ -571,7 +572,7 @@ import { version as sdkVersion } from '../../package.json';
 
     EndUserConsent.addEventListener(
       'end-user-consent.totp.denied.contact-methods-outdated',
-      (event) => {
+      () => {
         (referenceWindow.parent || referenceWindow).postMessage(
           'SmileIdentity::ConsentDenied::TOTP::ContactMethodsOutdated',
           '*',
@@ -881,6 +882,7 @@ import { version as sdkVersion } from '../../package.json';
         ...partner_params,
         job_type: 1,
       },
+      metadata: getMetadata(),
     };
 
     const URL = `${getEndpoint(config.environment)}/upload`;