+++ packages/web-components/src/components/selfie/src/SelfieCaptureScreens.js
@@ -7,12 +7,25 @@ import packageJson from '../../../../package.json';
 
 const COMPONENTS_VERSION = packageJson.version;
 
+const smartCameraWeb = document.querySelector('smart-camera-web');
+
 async function getPermissions(captureScreen, facingMode = 'user') {
   try {
-    await SmartCamera.getMedia({
+    const stream = await SmartCamera.getMedia({
       audio: false,
       video: { facingMode },
     });
+    const devices = await navigator.mediaDevices.enumerateDevices();
+    const videoDevice = devices.find(
+      (device) =>
+        device.kind === 'videoinput' &&
+        stream.getVideoTracks()[0].getSettings().deviceId === device.deviceId,
+    );
+    smartCameraWeb?.dispatchEvent(
+      new CustomEvent('metadata.camera-name', {
+        detail: { cameraName: videoDevice?.label },
+      }),
+    );
     captureScreen.removeAttribute('data-camera-error');
     captureScreen.setAttribute('data-camera-ready', true);
   } catch (error) {
@@ -28,6 +41,7 @@ class SelfieCaptureScreens extends HTMLElement {
   constructor() {
     super();
     this.activeScreen = null;
+    smartCameraWeb?.dispatchEvent(new CustomEvent('metadata.initialize'));
   }
 
   connectedCallback() {
@@ -89,6 +103,19 @@ class SelfieCaptureScreens extends HTMLElement {
         await getPermissions(this.selfieCapture, this.getAgentMode()).then(() =>
           this.setActiveScreen(this.selfieCapture),
         );
+        smartCameraWeb?.dispatchEvent(
+          new CustomEvent('metadata.selfie-capture-start'),
+        );
+        smartCameraWeb?.dispatchEvent(
+          new CustomEvent('metadata.selfie-origin', {
+            detail: {
+              imageOrigin: {
+                environment: 'back_camera',
+                user: 'front_camera',
+              }[this.getAgentMode()],
+            },
+          }),
+        );
       },
     );
     this.selfieInstruction.addEventListener(
@@ -107,6 +134,9 @@ class SelfieCaptureScreens extends HTMLElement {
     });
 
     this.selfieCapture.addEventListener('selfie-capture.publish', (event) => {
+      smartCameraWeb?.dispatchEvent(
+        new CustomEvent('metadata.selfie-capture-end'),
+      );
       this.selfieReview.setAttribute('data-image', event.detail.referenceImage);
       this._data.images = event.detail.images;
       SmartCamera.stopMedia();
@@ -127,6 +157,9 @@ class SelfieCaptureScreens extends HTMLElement {
     this.selfieReview.addEventListener(
       'selfie-capture-review.rejected',
       async () => {
+        smartCameraWeb?.dispatchEvent(
+          new CustomEvent('metadata.selfie-capture-retry'),
+        );
         this.selfieReview.removeAttribute('data-image');
         this._data.images = [];
         if (this.hideInstructions) {
@@ -194,7 +227,10 @@ class SelfieCaptureScreens extends HTMLElement {
   }
 
   get hideBack() {
-    return this.hasAttribute('hide-back-to-host') ? 'hide-back' : '';
+    return this.hasAttribute('hide-back-to-host') ||
+      this.hasAttribute('hide-back')
+      ? 'hide-back'
+      : '';
   }
 
   get disableImageTests() {