+++ previews/app/routes/products.$productName.tsx
@@ -124,7 +124,8 @@ export default function Product() {
           <form onSubmit={handleSubmit}>
             <fieldset disabled={isGettingToken}>
               <legend>
-                <h1>{product.label}</h1>
+                <h1>{product.label}</h1>{' '}
+                <span>{appStage ? `Previewing ${appStage}` : ''}</span>
               </legend>
 
               <input type="hidden" name="product" value={product.name} />
@@ -169,7 +170,7 @@ export default function Product() {
 
           <script
             src={`https://cdn.smileidentity.com/inline/${
-              appStage.includes('preview/') ? appStage : 'v1'
+              appStage === 'main' ? 'v1' : `preview-${appStage}`
             }/js/script.min.js`}
           ></script>
         </>